IMPORTANT CHANGES
The repository is now private, so we must be extra careful. I've created a .gitignore file to ignore some files like the new credentials.env file which stores secretly all the database credentials. I think that I'm the only one with the file in my computer at the moment so before you do any changes talk to me and I will tell you what to do.