API bug fixed
The API got the user from the PC's user environmental variable, so I've changed the name of the variable in the .env file to DBUSER to make it different. The .gitignore has been updated because I didn't want to push the file that I've created for myself to debug the API. I've also found two more bugs in the API but they have an easy fix, if you read the file you'll find two FIXME comments explaining more.