Started coding a way of navigating the CC
I started implementating a way of navigating the control center. Basically it consists of a DB row selection. It should also be useful for .txt files that aren't uploaded in the DB. I've also created a log file to do some testing, it is temporal. **If you want to read the script you will find some of my notes as well as the things that need to be implementated. Feel free to add any of these features.**